_ML/requirements_GPU_pip.txt
absl-py==1.2.0
astor==0.8.1
cachetools==4.2.4
certifi==2022.6.15.1
charset-normalizer==2.1.1
cloudpickle==2.2.0
cycler==0.11.0
cytoolz==0.12.0
dask==2022.2.0
decorator==5.1.1
fonttools==4.37.1
fsspec==2022.8.2
gast==0.2.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.1
h5py==3.7.0
idna==3.3
imagecodecs==2021.11.20
imageio==2.21.3
imgaug==0.3.0
importlib-metadata==4.12.0
joblib==1.1.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
locket==1.0.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
munkres==1.1.4
natsort==7.0.1
networkx==2.5
numpy==1.21.6
oauthlib==3.2.1
opencv-python==4.6.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.0.5
partd==1.3.0
Pillow==9.2.0
pip==22.2.2
ply==3.11
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-sip==12.11.0
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.3.0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.10.1
setuptools==65.3.0
Shapely==1.8.4
sip==6.6.2
six==1.16.0
tensorboard==2.1.1
tensorflow-estimator==2.1.0
tensorflow-gpu==2.1.0
termcolor==2.0.1
threadpoolctl==3.1.0
tifffile==2021.11.2
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.47.0
typing_extensions==4.3.0
unicodedata2==14.0.0
urllib3==1.26.12
Werkzeug==2.2.2
wheel==0.37.1
wrapt==1.14.1
zipp==3.8.1
